Implement provider for WarpStream Virtual Clusters
This repository is based on the Terraform Plugin Framework template repoistory
on GitHub: https://github.com/hashicorp/terraform-provider-scaffolding-framework

The Go Client for interacting with the WarpStream API is implemented inside
folder `/internal/provider/api/`. Any change in the WarpStream API will have
to be reflected in this HTTP Go Client.

The main terraform resource implemented by this first version of the provider
is `warpstream_virtual_cluster`. A corresponding data source is also available.