handle resources not found except topics
This modifies logic on all resources to properly handle deletion
outside of terraform.

If a resource is deleted a plan will no longer fail when a 404 is returned.
If a resource is deleted a destroy will no longer fail when a 404 is returned.

Topics have no been modified because the API returns a 500 instead of a 404.
Topics will get updated once the API is fixed.