Implement deletion protection for virtual clusters.
Adds a new "enable_deletion_protection" field to
the configuration of a virtual cluster.

This field is used to prevent accidental deletion of a
virtual cluster.

When set to true, it's impossible to delete the
virtual cluster in the UI, or the API, or terraform.

```
resource "warpstream_virtual_cluster" "test" {
  name = "vcn_test_"
  configuration = {
    enable_deletion_protection = true
  }
}
```